telegram_bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
asyncio
aiohttp==3.9.1
python-dotenv==1.0.0
schedule==1.2.0
datetime
logging